Better randomness
I have noticed an error in my series generation.
According to the book there are 2 dice rolls for 3 and 4 series and
only one for 2 and 5 series. So they get generated more often.

Also there should be a reverse in exercises every week, but I haven't
finished implementing that idea yet - too much depends on the index
instead of the selected exercise.